fix: report setup provisioning user agents
The plugin User-Agent rollout needs every server-bound plugin request to identify the plugin name and version, including first-time setup provisioning calls.

Add User-Agent headers to Codex and OpenCode API key provisioning, share the OpenCode user-agent helper across runtime and setup code, and cover both setup paths in tests.
